+++ src/stories/header.css
@@ -21,6 +21,7 @@ h1 {
   vertical-align: top;
 }
 
-button + button {
+/* 意図せぬCSSが当たるのでコメントアウト */
+/* button + button {
   margin-left: 10px;
-}
+} */
